feat: remove the tools module (and base exceptions module) and put context container logic in the nodes module.
docs: update docs to reflect refactored context container call
