+++ docs/api/pytest_houdini.rst
@@ -12,26 +12,10 @@ Subpackages
 Submodules
 ----------
 
-pytest\_houdini.exceptions module
----------------------------------
-
-.. automodule:: pytest_houdini.exceptions
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
 pytest\_houdini.plugin module
 -----------------------------
 
 .. automodule:: pytest_houdini.plugin
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-pytest\_houdini.tools module
-----------------------------
-
-.. automodule:: pytest_houdini.tools
    :members:
    :undoc-members:
    :show-inheritance:
